website_risk_checker/requirements.txt
streamlit
validators
requests
python-whois
